faq 완료

+++ src/main/java/yiu/aisl/devTogether/service/NoticeService.java
@@ -1,6 +1,5 @@
 package yiu.aisl.devTogether.service;
 
-import ch.qos.logback.classic.Logger;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,7 +12,6 @@ import yiu.aisl.devTogether.exception.CustomException;
 import yiu.aisl.devTogether.exception.ErrorCode;
 import yiu.aisl.devTogether.repository.NoticeRepository;
 
-import yiu.aisl.devTogether.domain.User;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -23,7 +21,7 @@ import java.util.Optional;
 @Transactional
 public class NoticeService {
     private final NoticeRepository noticeRepository;
-    private Logger log;
+
 
     //공지사항 전체조회
     public List<NoticeResponseDto> getList() throws  Exception{
@@ -33,29 +31,21 @@ public class NoticeService {
         return getListDTO;
     }
 
+
     //공지사항 등록
-    public Boolean create(NoticeRequestDto.CreateDTO request, User user) {
-        // 400 데이터 미입력
-        if (request.getTitle() == null || request.getContents() == null
-                || request.getNoticeCategory() == null || request.getRoleCategory() == null)
-            throw new CustomException(ErrorCode.INSUFFICIENT_DATA);
-
-
-
-        NoticeCategory noticeCategory = NoticeCategory.fromInt(request.getNoticeCategory());
+    public Boolean create(NoticeRequestDto.CreateDTO request) {
+        NoticeCategory noticeCategory = NoticeCategory.fromInt(request.getNoticeCategory());    //열거형 상수
         RoleCategory roleCategory = RoleCategory.fromInt(request.getRoleCategory());
-        int roleCategoryValue = request.getRoleCategory();
-
-        // 관리자 권한 확인
-        if (roleCategoryValue != RoleCategory.MANAGER.getValue()) {
+        // 400 - 데이터 미입력
+        if (request.getTitle() == null || request.getContents() == null
+                || request.getNoticeCategory() == null || request.getRoleCategory() == null
+                || request.getFile() == null)
+            throw new CustomException(ErrorCode.INSUFFICIENT_DATA);
+        // 403 - 권한 없음
+        if (roleCategory != RoleCategory.MANAGER) {
             throw new CustomException(ErrorCode.NO_AUTH);
         }
-
         try {
-
-
-
-            // Notice 객체 생성 및 저장
             Notice notice = Notice.builder()
                     .roleCategory(roleCategory)
                     .title(request.getTitle())
@@ -63,7 +53,6 @@ public class NoticeService {
                     .file(request.getFile())
                     .noticeCategory(noticeCategory)
                     .build();
-
             noticeRepository.save(notice);
         } catch (Exception e) {
             throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
@@ -74,9 +63,16 @@ public class NoticeService {
 
     //공지사항 삭제
     public Boolean delete(NoticeRequestDto.DeleteDTO request) {
-
+        RoleCategory roleCategory = RoleCategory.fromInt(request.getRoleCategory());
         // 404 - id 없음
         Notice notice = findBynNoticeId(request.getNoticeId());
+        if(notice == null){
+            throw new CustomException(ErrorCode.NOT_EXIST_ID);
+        }
+        // 403 - 권한 없음
+        if(roleCategory != RoleCategory.MANAGER){
+            throw new CustomException(ErrorCode.NO_AUTH);
+        }
         try{
             //deleteByNoticeId로 하면 SQL에서 NoticeId 프로퍼티를 인식하지못함
             noticeRepository.deleteById(request.getNoticeId());
@@ -85,6 +81,7 @@ public class NoticeService {
         catch (Exception e){
             throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
         }
+
     }
 
 
@@ -92,37 +89,32 @@ public class NoticeService {
 
     // 공지사항 수정
     public Boolean update(NoticeRequestDto.UpdateDTO request) {
-        // 400 데이터 미입력
-        if (request.getTitle() == null || request.getContents() == null || request.getNoticeCategory() == null
+        NoticeCategory noticeCategory = NoticeCategory.fromInt(request.getNoticeCategory());
+        RoleCategory roleCategory = RoleCategory.fromInt(request.getRoleCategory());
+        // 400 - 데이터 미입력
+        if (request.getTitle() == null || request.getContents() == null
+                || request.getNoticeCategory() == null || request.getRoleCategory() == null
                 || request.getNoticeId() == null|| request.getFile() == null) {
             throw new CustomException(ErrorCode.INSUFFICIENT_DATA);
         }
+        // 403 - 권한 없음
+        if (roleCategory != RoleCategory.MANAGER) {
+            throw new CustomException(ErrorCode.NO_AUTH);
+        }
         try {
-
-            NoticeCategory noticeCategory = NoticeCategory.fromInt(request.getNoticeCategory());
-
-
             Optional<Notice> modifyNotice = noticeRepository.findByNoticeId(request.getNoticeId());
-            if (modifyNotice.isEmpty()) {
-                throw new CustomException(ErrorCode.NOT_EXIST_ID);
-            }
-
-
             Notice modifiedNotice = modifyNotice.get();
+            modifiedNotice.setRoleCategory(roleCategory);
             modifiedNotice.setTitle(request.getTitle());
             modifiedNotice.setContents(request.getContents());
             modifiedNotice.setNoticeCategory(noticeCategory);
             modifiedNotice.setFile(request.getFile());
-
-
             noticeRepository.save(modifiedNotice);
         }    catch (Exception e){
             throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
         }
         return true;
     }
-
-
     private Notice findBynNoticeId(Long noticeId) {
         return noticeRepository.findByNoticeId(noticeId)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_EXIST_MEMBER));
